test: fix stale dtype-rejection assertions for tensor-mlx/tensor-webgpu
No source changes needed here -- confirmed neither facade hardcodes
its own DType union (both import/re-export tensor-backend's). Just
test fixes surfaced by verifying against the widened backends:

- Stale regexes expecting the old bridge-level "no float64"/
  "cast(...)" wording now that #checkDtype's generic message differs,
  and one test that predated f64/i64/u8 being valid tensor-core dtypes
  at all (hostFromTensor no longer rejects them -- per-backend
  rejection happens at upload time instead).
- Conformance test filters excluding cases whose own dtype a given
  backend permanently can't support (f64 on MLX's GPU device; the
  WGSL-spec-capped dtypes on WebGPU) -- mirrors the same fix already
  made in backend-mlx/backend-webgpu's own conformance tests, since
  the shared harness only checks supports() for the outer f32/f16/bf16
  run dtype, not each case's own fixed dtype.
- One incidental type-only fix (write-buffer-offset.bun.ts) for a
  HostData union now including bigint arrays.

Verified: tensor-mlx 233/233, tensor-webgpu 259/260 (0 fail; the one
skip is a pre-existing "no Bun on PATH" environment gap).

+++ packages/tensor-webgpu/test/bun/write-buffer-offset.bun.ts
@@ -44,7 +44,7 @@ rt.write(viaWrite, false, view);
 const writeBytesOut = Array.from(new Float32Array(await rt.readBytes(viaWrite, 0, 16)));
 
 const t32 = await gpu.fromHost({ dtype: "f32", shape: [4], data: view });
-const tensorF32 = Array.from((await gpu.toHost(t32)).data);
+const tensorF32 = Array.from((await gpu.toHost(t32)).data as Float32Array);
 
 const expected = { f32: [4, 5, 6, 7], f16: [0x3c03, 0x3c04, 0x3c05, 0x3c06] };
 let tensorF16: number[] | null = null;
